Change exposure analysis icon. Seperate download and analysis logic
