+++ frontend/src/components/MapView/LeftPanel/layersPanel/MenuSwitch/SwitchItem/LayerDownloadOptions.tsx
@@ -10,13 +10,9 @@ import React, { useState } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import moment from 'moment';
 import { mapValues } from 'lodash';
-import BarChartIcon from '@material-ui/icons/BarChart';
 import GetAppIcon from '@material-ui/icons/GetApp';
 import {
   AdminLevelDataLayerProps,
-  AggregationOperations,
-  ExposedPopulationDefinition,
-  GeometryType,
   LayerType,
   WMSLayerProps,
 } from '../../../../../../config/types';
@@ -33,26 +29,15 @@ import {
 import { castObjectsArrayToCsv } from '../../../../../../utils/csv-utils';
 import { downloadGeotiff, Extent } from '../../../../Layers/raster-utils';
 import { useSafeTranslation } from '../../../../../../i18n';
-import {
-  analysisResultSelector,
-  clearAnalysisResult,
-  ExposedPopulationDispatchParams,
-  isExposureAnalysisLoadingSelector,
-  requestAndStoreExposedPopulation,
-  setCurrentDataDefinition,
-} from '../../../../../../context/analysisResultStateSlice';
-import { setTabValue } from '../../../../../../context/leftPanelStateSlice';
-import { TableKey } from '../../../../../../config/utils';
+import { isExposureAnalysisLoadingSelector } from '../../../../../../context/analysisResultStateSlice';
 
 function LayerDownloadOptions({
   layer,
   extent,
   selected,
-  exposure,
 }: LayerDownloadOptionsProps) {
   const { t } = useSafeTranslation();
   const dispatch = useDispatch();
-  const analysisResult = useSelector(analysisResultSelector);
 
   const [
     downloadMenuAnchorEl,
@@ -74,13 +59,6 @@ function LayerDownloadOptions({
 
   const handleDownloadMenuOpen = (event: React.MouseEvent<HTMLElement>) => {
     setDownloadMenuAnchorEl(event.currentTarget);
-  };
-
-  // Since the exposure analysis doesn't have predefined table in configurations
-  // and need to have a `TableKey` will use this util function to handle such case
-  // used timestamp to avoid any potential rare name collision
-  const generateUniqueTableKey = (activityName: string) => {
-    return `${activityName}_${Date.now()}`;
   };
 
   const getFilename = (): string => {
@@ -138,45 +116,6 @@ function LayerDownloadOptions({
     handleDownloadMenuClose();
   };
 
-  const handleExposureAnalysis = () => {
-    if (analysisResult) {
-      dispatch(clearAnalysisResult());
-    }
-
-    if (!layer.id || !extent || !exposure) {
-      return;
-    }
-
-    if (!selectedDate) {
-      throw new Error('Date must be given to run analysis');
-    }
-
-    const hazardLayer =
-      layer.type === 'wms' && layer.geometry === GeometryType.Polygon
-        ? { wfsLayerId: layer.id }
-        : { maskLayerId: layer.id };
-
-    const params: ExposedPopulationDispatchParams = {
-      exposure,
-      date: selectedDate,
-      statistic: AggregationOperations.Sum,
-      extent,
-      ...hazardLayer,
-    };
-
-    dispatch(requestAndStoreExposedPopulation(params));
-    dispatch(
-      setCurrentDataDefinition({
-        id: generateUniqueTableKey('exposure_analysis') as TableKey,
-        title: analysisResult?.getTitle(t) || '',
-        table: '',
-        legendText: t(analysisResult?.legendText || ''),
-      }),
-    );
-    handleDownloadMenuClose();
-    dispatch(setTabValue(2));
-  };
-
   const shouldShowDownloadButton =
     layer.type === 'admin_level_data' ||
     (layer.type === 'wms' &&
@@ -184,14 +123,6 @@ function LayerDownloadOptions({
 
   return (
     <>
-      {exposure && (
-        <Tooltip title={t('Exposure Analysis') ?? ''}>
-          <IconButton disabled={!selected} onClick={handleExposureAnalysis}>
-            <BarChartIcon />
-          </IconButton>
-        </Tooltip>
-      )}
-
       {shouldShowDownloadButton && (
         <Tooltip title="Download">
           <IconButton
@@ -238,7 +169,6 @@ interface LayerDownloadOptionsProps {
   layer: LayerType;
   extent: Extent | undefined;
   selected: boolean;
-  exposure: ExposedPopulationDefinition | undefined;
 }
 
 export default LayerDownloadOptions;
