+++ frontend/src/components/MapView/LeftPanel/layersPanel/MenuSwitch/SwitchItem/index.tsx
@@ -39,6 +39,7 @@ import { getUrlKey, useUrlHistory } from '../../../../../../utils/url-utils';
 import { handleChangeOpacity } from '../../../../Legends/handleChangeOpacity';
 import { Extent } from '../../../../Layers/raster-utils';
 import LayerDownloadOptions from './LayerDownloadOptions';
+import ExposureAnalysisOption from './ExposureAnalysisOption';
 
 /**
  * Returns layer identifier used to perform exposure analysis.
@@ -221,11 +222,18 @@ function SwitchItem({ classes, layer, extent }: SwitchItemProps) {
             <OpacityIcon />
           </IconButton>
         </Tooltip>
+        {exposure && (
+          <ExposureAnalysisOption
+            layer={layer}
+            extent={extent}
+            selected={selected}
+            exposure={exposure}
+          />
+        )}
         <LayerDownloadOptions
           layer={layer}
           extent={extent}
           selected={selected}
-          exposure={exposure}
         />
       </Box>
       {selected && isOpacitySelected && (
